some fine tuning

+++ bot.py
@@ -25,26 +25,26 @@ keyboard2 = InlineKeyboardMarkup(inline_keyboard=[
     [InlineKeyboardButton('saved3', callback_data='saved3'),
      InlineKeyboardButton('lazada', callback_data='lazada'),
      InlineKeyboardButton('family', callback_data='family')],
+    [InlineKeyboardButton('Done', callback_data='cancel')],
 ])
 
 
 # one handler for debug them all
 @dp.message_handler(lambda message: message.text and 'getinfo' in message.text.lower())
 async def start_process(message: types.Message):
-    await message.reply(message)
-
+    # await message.reply(message)
+    await bot.send_message(chat_id=LOCAL, text=message)
 
 # message handlers starts here
-@dp.message_handler(lambda message: message.text and '#notes' in message.text.lower())
-async def start_process(message: types.Message):
-    await message.send_copy(chat_id=NOTES, disable_notification=True)
-    await message.delete()
-
-
 @dp.message_handler(lambda message: message.text and '#saved' in message.text.lower())
 async def send_welcome(message: types.Message):
     await message.send_copy(chat_id=SAVED2, disable_notification=True)
     await message.delete()
+
+
+@dp.message_handler(lambda message: message.text and '#' in message.text.lower())
+async def start_process(message: types.Message):
+    await message.send_copy(chat_id=NOTES, disable_notification=True)
 
 
 @dp.message_handler(commands=['start', 'help'], chat_id=LOCAL)
@@ -58,14 +58,22 @@ async def send_dice(message: types.Message):
 
 
 # main handler
-@dp.message_handler(content_types=ContentType.ANY, chat_id=LOCAL)
+# @dp.message_handler(content_types=ContentType.ANY, chat_id=LOCAL, user_id=LOCAL)
+@dp.message_handler(content_types=ContentType.ANY)
 async def other(message: types.Message):
-    if message.forward_date:
-        message_timestamp = message.forward_date.strftime('%Y-%m-%d %H:%M:%S')
-        message_data = f'{message.forward_from_chat.title} at {message_timestamp}'
-        await message.answer(message_data)
-    await message.send_copy(chat_id=LOCAL, reply_markup=keyboard2, disable_notification=True)
-    await message.delete()
+    if message.chat.id > 0:
+        chat_id = message.chat.id
+        if message.sticker:
+            await message.answer(message.sticker.thumb.file_id)
+        if message.forward_date:
+            message_timestamp = message.forward_date.strftime('%Y-%m-%d %H:%M:%S')
+            if message.forward_from_chat:
+                message_data = f'{message.forward_from_chat.title} at {message_timestamp}'
+            else:
+                message_data = f'{message.forward_from.username} at {message_timestamp}'
+            await message.answer(message_data)
+        await message.send_copy(chat_id=chat_id, reply_markup=keyboard2, disable_notification=True)
+        await message.delete()
 
 
 # Query handlers
@@ -107,6 +115,12 @@ async def process_callback_button1(callback_query: types.CallbackQuery):
 @dp.callback_query_handler(text='saved3')
 async def process_callback_button1(callback_query: types.CallbackQuery):
     await callback_query.message.send_copy(chat_id=SAVED3, reply_markup=keyboard, disable_notification=True)
+    await callback_query.message.delete()
+
+
+@dp.callback_query_handler(text='cancel')
+async def process_callback_button1(callback_query: types.CallbackQuery):
+    await callback_query.message.answer('Галя, у нас отмена!')
     await callback_query.message.delete()
 
 if __name__ == '__main__':
